Place footer in all tab bars + refactoring
